ooDialog -   * Fill out the MonthCalendar class.   * Implement the ground work to directly invoke the Rexx dialog's methods     for Windows event messages.  This allows the programmer to directly     reply to a Windows message.   * Clean up some of the handling of failures for dialog objects to ensure     that an incorrectly initialized dialog won't be used by the programmer.     In particular when the MAX dialog count is reached.
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@5481 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ extensions/platform/windows/oodialog/oodCommon.cpp
@@ -295,7 +295,7 @@ static bool isStaticID(RexxMethodContext *c, RexxObjectPtr id)
  *           being preserved.  Use oodSafeResolveID() for those cases.
  */
 uint32_t oodResolveSymbolicID(RexxMethodContext *context, RexxObjectPtr oodObj, RexxObjectPtr id,
-                              int argPosObj, int argPosID)
+                              int argPosObj, size_t argPosID)
 {
     uint32_t result = OOD_ID_EXCEPTION;
 
@@ -369,7 +369,7 @@ done_out:
  *        an exception is raised, it is cleared and false returned.
  */
 bool oodSafeResolveID(uint32_t *pID, RexxMethodContext *context, RexxObjectPtr oodObj, RexxObjectPtr id,
-                   int argPosObj, int argPosID)
+                   int argPosObj, size_t argPosID)
 {
     uint32_t tmp = oodResolveSymbolicID(context, oodObj, id, argPosObj, argPosID);
     if ( tmp == OOD_ID_EXCEPTION )
